Remove CoreActivity
