+++ app/src/main/java/io/horizontalsystems/bankwallet/core/BaseActivity.kt
@@ -6,13 +6,13 @@ import android.os.Bundle
 import android.view.View
 import android.view.WindowManager
 import android.view.inputmethod.InputMethodManager
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.WindowCompat
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
-import io.horizontalsystems.core.CoreActivity
 import io.horizontalsystems.core.hideKeyboard
 
-abstract class BaseActivity : CoreActivity(), NavController.OnDestinationChangedListener {
+abstract class BaseActivity : AppCompatActivity(), NavController.OnDestinationChangedListener {
 
     protected fun hideSoftKeyboard() {
         getSystemService(InputMethodManager::class.java)?.hideSoftInputFromWindow(currentFocus?.windowToken, 0)
